additional restructuring
