+++ src/domain/space/admin/SpaceCommunication/OpportunityCommunicationsPage.tsx
@@ -1,5 +1,7 @@
 import { PropsWithChildren } from 'react';
-import CommunityUpdatesPage, { CommunityUpdatesPageProps } from '@/domain/space/pages/CommunityUpdatesPage';
+import CommunityUpdatesPage, {
+  CommunityUpdatesPageProps,
+} from '@/domain/space/admin/SpaceCommunication/CommunityUpdatesPage';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
 import SubspaceSettingsLayout from '@/domain/space/routing/toReviewAdmin/SubspaceSettingsLayout';
