+++ src/domain/space/admin/SpaceCommunication/SubspaceCommunicationsPage.tsx
@@ -1,7 +1,9 @@
 import React, { FC } from 'react';
 import { SettingsSection } from '@/domain/platform/admin/layout/EntitySettingsLayout/SettingsSection';
 import { SettingsPageProps } from '@/domain/platform/admin/layout/EntitySettingsLayout/types';
-import CommunityUpdatesPage, { CommunityUpdatesPageProps } from '@/domain/space/pages/CommunityUpdatesPage';
+import CommunityUpdatesPage, {
+  CommunityUpdatesPageProps,
+} from '@/domain/space/admin/SpaceCommunication/CommunityUpdatesPage';
 import SubspaceSettingsLayout from '@/domain/space/routing/toReviewAdmin/SubspaceSettingsLayout';
 
 interface SubspaceCommunicationsPageProps extends SettingsPageProps, CommunityUpdatesPageProps {}
